requirements.txt
--extra-index-url https://www.paddlepaddle.org.cn/packages/stable/cu126/
--extra-index-url https://download.pytorch.org/whl/cu126
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aistudio_sdk==0.3.5
annotated-types==0.7.0
anyio==4.10.0
astor==0.8.1
attrs==25.3.0
bce-python-sdk==0.9.42
beautifulsoup4==4.14.2
boto3==1.35.23
botocore==1.35.23
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.3
cryptography==46.0.1
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
einops==0.8.1
et_xmlfile==2.0.0
filelock==3.20.1
fire==0.7.1
frozenlist==1.7.0
fsspec==2025.7.0
ftfy==6.3.1
future==1.0.0
fuzzywuzzy==0.18.0
GPUtil==1.4.0
greenlet==3.2.4
grpcio==1.75.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
imagesize==1.4.1
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
joblib==1.5.1
jpype1==1.6.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
lxml==6.0.2
Markdown==3.9
MarkupSafe==3.0.2
marshmallow==3.26.2
modelscope==1.29.0
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.5
numpy==2.2.6
openai==1.109.1
opencv-contrib-python==4.10.0.84
opencv-python>=4.11
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.3
packaging==25.0
paddleocr==3.3.2
paddlepaddle-gpu==2.6.2
paddlex==3.3.12
pandas==2.3.1
pdf2image==1.17.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==11.3.0
polars==1.33.1
premailer==3.10.0
prettytable==3.16.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pybboxes==0.1.6
pyclipper==1.3.0.post6
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.11.0
pydantic_core==2.33.2
PyMuPDF==1.26.3
pyparsing==3.2.3
pypdf==6.4.0
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.18
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
s3transfer==0.10.4
safetensors==0.6.2
sahi>=0.11.34
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
sentence-transformers==5.1.0
sentry-sdk==2.38.0
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.49.1
sympy>=1.13.3
tabula-py==2.10.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.3
triton==3.4.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.11.0
ultralytics==8.0.43
ultralytics-thop==2.0.17
ultralyticsplus==0.0.28
urllib3==2.6.0
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.4
wheel==0.45.1
yarl==1.20.1
zstandard==0.25.0
boto3
aws-lambda-powertools==3.23.0